feat: Ensure MkDocs and Allure Reports Always Deploy
This commit updates our GitHub Actions workflow to ensure that the
'MkDocs' and 'Allure Reports' jobs always run, even if the 'Build' job
fails. Previously, a failure in 'Build' would prevent these jobs from
running.

This change is important because it ensures that we always generate and
publish our documentation and test reports, providing critical insight
into the state of the project, regardless of whether the full CI pipeline
passes.

Please note, this change does not modify the dependencies between jobs.
The 'MkDocs' and 'Allure Reports' jobs will still wait for the 'Build' job
to complete before they start.