fix: Resolve file access issue in Allure reporting step and refine workflow structure
This commit addresses an issue with file permissions during the Allure reporting step of our CI pipeline. It also refines the structure of the entire workflow for better readability and maintainability.

Changes implemented in this commit:
- Add command to clean allure-history directory before writing new report, avoiding file permission conflicts.
- Add allure-history/* to .gitignore to prevent tracking of Allure history files, which could cause file access issues.
- Update use of 'peaceiris/actions-gh-pages' action to version v4.1.1 for enhanced stability and performance.
- Refine structure of steps in the 'deploy_allure' job, ensuring the generation and publication of the Allure report are consistently successful.
- Modify the condition checks using 'always()' to make sure every step is executed irrespective of previous failures.

Please note: This update aims to improve the robustness and fault tolerance of our CI pipeline. Ensure that GitHub Pages is enabled in the repository settings for the deployment to succeed.