fix: Update version of peaceiris/actions-gh-pages action in CI pipeline
This commit updates the version of the peaceiris/actions-gh-pages GitHub Action used in our CI pipeline. The update resolves an error that occurred due to the use of an unresolvable version of the action.

Changes implemented in this commit:
- Updated the version of peaceiris/actions-gh-pages from v4.1.1 to v3.9.3.

Please note: It's important to keep GitHub Actions updated to the latest resolvable versions to ensure the smooth operation of our CI pipeline.