feat(testing): add more logging to afterEach block
This commit introduces additional logging to the afterEach block in our testing suite. This change will help us better understand the execution time of our tests, particularly when they fail.

The added logging statements record the start and end times of the afterEach block, as well as whether the test status check resulted in